sensor backend restored to http based communication. Parses data from sensor module
